requirements.txt
asn1crypto==1.4.0
azure-common==1.1.25
azure-core==1.8.2
azure-storage-blob==12.5.0
boto3==1.15.18
botocore==1.18.18
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
croniter==0.3.37
cryptography==2.9.2
dask==2021.4.1
distributed==2021.4.1
docker==5.0.0
fsspec==2021.4.0
HeapDict==1.0.1
idna==2.10
isodate==0.6.0
jmespath==0.10.0
locket==0.2.1
marshmallow==3.11.1
marshmallow-oneofschema==2.1.0
msgpack==1.0.2
msrest==0.6.19
mypy-extensions==0.4.3
natsort==7.1.1
oauthlib==3.1.0
oscrypto==1.2.1
partd==1.2.0
pendulum==2.1.2
prefect==0.14.17
psutil==5.8.0
pycparser==2.20
pycryptodomex==3.9.8
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-box==5.3.0
python-dateutil==2.8.1
python-slugify==5.0.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.3.4
sortedcontainers==2.3.0
tabulate==0.8.9
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
urllib3==1.25.11
websocket-client==0.59.0
zict==2.0.0
